+++ states_rebuilder_package/test/injected_a_to_z/17-persiting_the_state_test.dart
@@ -52,7 +52,7 @@ class App extends StatelessWidget {
 void main() async {
   //Inject a mocked implementation of ILocalStorage
   //it return a store of type Mao
-  final store = (await RM.localStorageInitializerMock()).store;
+  final store = (await RM.storageInitializerMock()).store;
   setUp(() {
     store.clear();
   });